requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
google-generativeai
gunicorn
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.1.20
langchain-core==0.3.29
langchain-google-genai
langchain-ollama==0.1.0
langsmith==0.2.10
MarkupSafe==3.0.2
matplotlib==3.10.0
nltk==3.8.1
numpy==2.2.1
openpyxl
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.16.0
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
textblob==0.17.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wordcloud==1.9.4
xlrd
